+++ CHANGELOG.md
@@ -26,6 +26,7 @@ To enable, set `do_coalesced_comms=true` in the `<parthenon/mesh>` block of the 
 by adding an optional string as last argument to any `ParameterInput` `Get` or `GetOrAdd` call.
 
 ### Added (new features/APIs/variables/...)
+- [[PR 162]](https://github.com/parthenon-hpc-lab/athenapk/pull/162) Add pgen for cloud shattering setup
 - [[PR 158]](https://github.com/parthenon-hpc-lab/athenapk/pull/158) Update particle id handling (now automated `uint64`). Extend particle history lookback in turbulence pgen and include in turbulence test
 - [[PR 157]](https://github.com/parthenon-hpc-lab/athenapk/pull/157) Support injection of blobs with density/temp contrast in turbulence simulations